fix(user): 总爬升改整数显示（int 而非 float）
Tim 真用回归拍：73205.3 也不要 .3 / 直接 73205。
service: int(round(elevation_gain)) / schema: float → int / 配套防 JSON .0 尾巴。
"只修改这一处" 字段语义一处 / 跨 service+schema 2 行。

+++ app/user/schemas.py
@@ -106,7 +106,7 @@ class StatsResponse(BaseModel):
     period: str
     distance: float
     rides: int
-    elevation_gain: float
+    elevation_gain: int  # 米 / Tim 2026-05-11 拍整数显示 / int 而非 float 防 JSON 73205.0 .0 尾巴
     duration: int
     weekly_goal: float
     goal_percent: int
